+++ samples/framework/renderer.h
@@ -93,6 +93,17 @@ class Renderer {
                            const ozz::math::Float4x4& _transform,
                            bool _draw_joints = true) = 0;
 
+  // Renders points.
+  // _sizes and _colors must be either of ize 1 or equal to _positions' size.
+  // If _screen_space is true, then points size is fixed in screen-space,
+  // otherwise it changes with screen depth.
+  virtual bool DrawPointsIm(
+      const ozz::span<const ozz::math::Float3>& _positions,
+      const ozz::span<const float>& _sizes,
+      const ozz::span<const Color>& _colors,
+      const ozz::math::Float4x4& _transform, bool _round = true,
+      bool _screen_space = false) = 0;
+
   // Renders a box at a specified location.
   // The 2 slots of _colors array respectively defines color of the filled
   // faces and color of the box outlines.